\name{ozymandiasData}
\alias{ozymandiasData}
\docType{data}
\title{27k and 450k data for the ozymandias package example}
\description{
This package contains 450k and 27k titrated mixture sample IDATs 
for testing ozymandias' functionality as a preprocessing pipeline.
}
\examples{

## demo/example.R in ozymandias

setwd("~")
library(ozymandias)
message("finish this")

}
\keyword{datasets}
